requirements.txt
aiohttp
asyncio
discord.py
python-dotenv
requests
pyyaml == 5.3.1
sqlite3
